code review fixes: rename to pass-through, tests, lint
